+++ CHANGELOG.md
@@ -13,6 +13,8 @@ All notable changes to this project will be documented in this file.
 
 - Removed partitioning from direct payments pipeline outputs.
 
+- Calculate metrics function to now include proportion of estimates within 10 and 25 filled posts of the model value.
+
 - Added row_id and expanded_id to the ind cqc columns class as
   id_per_locationid_import_date and id_per_locationid_import_date_job_role respectively and
   updated references to the objects.